# backend/openapi.yaml
openapi: 3.0.0
info:
  title: PropertyPilot Backend API
  description: Real estate investment analysis platform with financial calculations, property management, and reporting.
  version: 1.0.0
  contact:
    name: PropertyPilot Support
    email: support@propertypilot.com
  license:
    name: MIT
servers:
  - url: https://api.propertypilot.com
    description: Production server
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/scenarios/{scenarioId}/analyze:
    post:
      tags:
        - Scenarios
      summary: Analyze scenario
      description: Calculate financial metrics for a scenario
      security:
        - SessionAuth: []
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      scenario:
                        $ref: '#/components/schemas/DealScenarioAnalyzed'
...
  /api/alerts:
    get:
      tags:
        - Alerts
      summary: List alerts
      description: Get all alerts for the authenticated user
      security:
        - SessionAuth: []
      parameters:
        - name: read
          in: query
          description: Filter by read status
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
  /api/alerts/{alertId}/read:
    put:
      tags:
        - Alerts
      summary: Mark alert as read
      description: Mark an alert as read
      security:
        - SessionAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      alert:
                        $ref: '#/components/schemas/Alert'
  /api/alerts/{alertId}:
    delete:
      tags:
        - Alerts
      summary: Delete alert
      description: Delete an alert
      security:
        - SessionAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Alert deleted successfully
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session-based authentication via httpOnly cookie
  schemas:
...
    DealScenario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
        name:
          type: string
        purchasePrice:
          type: number
          format: float
        rehabCost:
          type: number
          format: float
        holdingCosts:
          type: number
          format: float
        closingCosts:
          type: number
          format: float
        interestRate:
          type: number
          format: float
        holdTimeMonths:
          type: integer
        exitStrategy:
          type: string
          enum: [rent, airbnb, flip]
        monthlyRent:
          type: number
          format: float
          nullable: true
        occupancyRate:
          type: number
          format: float
          nullable: true
        dailyRate:
          type: number
          format: float
          nullable: true
        averageOccupancy:
          type: number
          format: float
          nullable: true
        salePrice:
          type: number
          format: float
          nullable: true
        sellingCosts:
          type: number
          format: float
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DealScenarioAnalyzed:
      allOf:
        - $ref: '#/components/schemas/DealScenario'
        - type: object
          properties:
            capRate:
              type: number
              format: float
            cashOnCash:
              type: number
              format: float
            roi:
              type: number
              format: float
            monthlyNOI:
              type: number
              format: float
            totalProfit:
              type: number
              format: float
            irr:
              type: number
              format: float
            npv:
              type: number
              format: float
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        message:
          type: string
        propertyId:
          type: string
          format: uuid
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...